Struct jsonwebtoken::DecodingKey

pub struct DecodingKey<'a> { /* private fields */ }

All the different kind of keys we can use to decode a JWT This key can be re-used so make sure you only initialize it once if you can for better performance

Implementations

impl<'a> DecodingKey<'a>

pub fn from_secret(secret: &'a [u8]) -> Self

If you’re using HMAC, use this.

pub fn from_base64_secret(secret: &str) -> Result<Self>

If you’re using HMAC with a base64 encoded, use this.

pub fn from_rsa_pem(key: &'a [u8]) -> Result<Self>

If you are loading a public RSA key in a PEM format, use this.

pub fn from_rsa_components(modulus: &'a str, exponent: &'a str) -> Self

If you have (n, e) RSA public key components, use this.

pub fn from_ec_pem(key: &'a [u8]) -> Result<Self>

If you have a ECDSA public key in PEM format, use this.

pub fn from_rsa_der(der: &'a [u8]) -> Self

If you know what you’re doing and have a RSA DER encoded public key, use this.

pub fn from_ec_der(der: &'a [u8]) -> Self

If you know what you’re doing and have a RSA EC encoded public key, use this.

pub fn into_static(self) -> DecodingKey<'static>

Convert self to DecodingKey<'static>.