jsonwebtoken

Module validation

Structs

BorrowedCowIfPossible 🔒

Usually #[serde(borrow)] on Cow enables deserializing with no allocations where possible (no escapes in the original str) but it does not work on e.g. HashSet<Cow<str>> We use this struct in this case.

ClaimsForValidation 🔒

Validation

Contains the various validations that are applied after decoding a JWT.

Enums

Audience 🔒

Issuer 🔒

TryParse 🔒

Functions

get_current_timestamp

Gets the current timestamp in the format expected by JWTs.

is_subset 🔒

numeric_type 🔒

validate 🔒