Enum jsonwebtoken::errors::ErrorKind

pub enum ErrorKind {
    Msg(String),
    Unspecified(Unspecified),
    Base64(DecodeError),
    Json(Error),
    Utf8(FromUtf8Error),
    InvalidToken,
    InvalidSignature,
    InvalidKey,
    ExpiredSignature,
    InvalidIssuer,
    InvalidAudience,
    InvalidSubject,
    InvalidIssuedAt,
    ImmatureSignature,
    InvalidAlgorithm,
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

Unspecified(Unspecified)

An error happened while signing/verifying a token with RSA

Base64(DecodeError)

An error happened while decoding some base64 text

Json(Error)

An error happened while serializing/deserializing JSON

Utf8(FromUtf8Error)

An error happened while trying to convert the result of base64 decoding to a String

InvalidToken

When a token doesn't have a valid JWT shape

InvalidSignature

When the signature doesn't match

InvalidKey

When the secret given is not a valid RSA key

ExpiredSignature

When a token’s exp claim indicates that it has expired

InvalidIssuer

When a token’s iss claim does not match the expected issuer

InvalidAudience

When a token’s aud claim does not match one of the expected audience values

InvalidSubject

When a token’s aud claim does not match one of the expected audience values

InvalidIssuedAt

When a token’s iat claim is in the future

ImmatureSignature

When a token’s nbf claim represents a time in the future

InvalidAlgorithm

When the algorithm in the header doesn't match the one passed to decode

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.