Module jsonwebtoken::jwk

Structs

CommonParameters

Common JWK parameters

EllipticCurveKeyParameters

Parameters for an Elliptic Curve Key

Jwk

JwkSet

A JWK set

OctetKeyPairParameters

Parameters for an Octet Key Pair

OctetKeyParameters

Parameters for an Octet Key

RSAKeyParameters

Parameters for a RSA Key

Enums

AlgorithmParameters

Algorithm specific parameters

EllipticCurve

Type of cryptographic curve used by a key. This is defined in RFC 7518 #7.6

EllipticCurveKeyType

Key type value for an Elliptic Curve Key. This single value enum is a workaround for Rust not supporting associated constants.

KeyOperations

Operations that the key is intended to be used for. This enum is serialized untagged

OctetKeyPairType

Key type value for an Octet Key Pair. This single value enum is a workaround for Rust not supporting associated constants.

OctetKeyType

Key type value for an Octet symmetric key. This single value enum is a workaround for Rust not supporting associated constants.

PublicKeyUse

The intended usage of the public KeyType. This enum is serialized untagged

RSAKeyType

Key type value for an RSA Key. This single value enum is a workaround for Rust not supporting associated constants.